requirements.txt
python-dotenv
rich
pydantic_ai
mcp
markdown
fastapi
uvicorn
mcp-use